Added some minor functions to the resources classes to help for code later on.
Added research class to allows for the research modifies to easily be located and used on resource collection types (is currently being drawn the the correct resource type tree;

Added the structures first abstract class named House which allows for now the creation and testing of a Structure that works. Yet to add anything to the structures with the exception of the GUI image box.